Update README.md
Conclusion
Oda revealed that he originally planned One Piece to last five years, and that he had already planned the ending. However, he found it would take longer than he had expected as Oda realized that he liked the story too much to end it in that period of time. In 2016, nineteen years after the start of serialization, the author said that the manga has reached 65% of the story he intends to tell. In July 2018, on the occasion of the twenty-first anniversary of One Piece, Oda said that the manga has reached 80% of the plot, while in January 2019, he said that One Piece is on its way to the conclusion, but that it could exceed the 100th volume.[19] In August 2019, Oda said that, according to his predictions, the manga will end between 2024 and 2025. However, Oda stated that the ending would be what he had decided in the beginning; he is committed to seeing it through. In a television special aired in Japan, Oda said he would be willing to change the ending if the fans were to be able to predict it.[5] In August 2020, Shueisha announced in the year's 35th issue of Weekly Shōnen Jump that One Piece was "headed toward the upcoming final saga. On January 4, 2021, One Piece reached its thousandth chapter. In June 2022, Oda announced that the manga would enter a one-month break to prepare for its 25th anniversary and its final saga, set to begin with the release of chapter 1054.